pyproject.toml
[project]
name = "mistral-ws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.0",
    "fastapi>=0.128.0",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "ollama>=0.6.1",
    "pymupdf>=1.26.7",
    "pypdf>=6.5.0",
    "python-docx>=1.2.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "transformers>=4.57.3",
    "uvicorn>=0.40.0",
]
